# Introduce cache_valid_time parameter