@cachedout 2016-07-26
I think this is good but there are a few minor items:

Could you please fix this lint error? https://jenkins.saltstack.com/job/PR/job/salt-pr-lint-n/3606/violations/file/salt/modules/aptpkg.py/

Also, let's add some documentation here so that people understand what this value does and how to set it.

After those two changes, I think we can get this in. Thanks.